feat: Termina configuração de autenticação e token e inclui configurações de exceptions em nossa aplicação

+++ src/main/resources/db/migration/V3__create_table_movie.sql
@@ -1,9 +1,9 @@
 CREATE TABLE movie (
-    id serial PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     title varchar(255) NOT NULL,
     description text,
     release_date date,
-    rating numeric,
+    rating double precision,
     created_at timestamp,
     updated_at timestamp
 );
